+++ packages/planets/src/planet/Planet.chunk.ts
@@ -1,3 +1,5 @@
+import { BufferAttribute, BufferGeometry, Float32BufferAttribute } from "three"
+import { MeshBVH } from "three-mesh-bvh"
 import { BuildChunkInitialParams, ChunkGeneratorProps } from "../chunk/types"
 import { fixEdgeSkirt } from "./chunk-helpers/fixEdgeSkirts"
 import { generateIndices } from "./chunk-helpers/generateIndices"
@@ -8,6 +10,8 @@ import { normalizeNormals } from "./chunk-helpers/normalizeNormals"
 export function buildPlanetChunk<D>(initialParams: BuildChunkInitialParams<D>) {
   const { heightGenerator, colorGenerator, terrainSplatGenerator } =
     initialParams
+
+  const geo = new BufferGeometry()
 
   return function runBuildChunk(params: ChunkGeneratorProps<D>) {
     const { resolution, origin, width, offset, radius, inverted } = params
@@ -56,12 +60,18 @@ export function buildPlanetChunk<D>(initialParams: BuildChunkInitialParams<D>) {
     coordsArray.set(coords, 0)
     indicesArray.set(indices, 0)
 
+    geo.setAttribute("position", new Float32BufferAttribute(positionsArray, 3))
+    geo.setIndex(new BufferAttribute(new Uint32Array(indicesArray), 1))
+    const bvh = new MeshBVH(geo)
+    const serializedBVH = MeshBVH.serialize(bvh)
+
     return {
       positions: positionsArray,
       colors: colorsArray,
       normals: normalsArray,
       uvs: coordsArray,
       indices: indicesArray,
+      bvh: serializedBVH,
     }
   }
 }