+++ packages/react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @hello-worlds/react
 
+## 0.0.22
+
+### Patch Changes
+
+- Updated dependencies
+  - @hello-worlds/planets@0.0.20
+
 ## 0.0.21
 
 ### Patch Changes